+++ projects/ccd-case-ui-toolkit/src/lib/shared/pipes/complex/index.ts
@@ -2,5 +2,5 @@ export * from './ccd-collection-table-value-case-fields.pipe';
 export * from './ccd-cyapage-label-filter.pipe';
 export * from './ccd-read-fields-filter.pipe';
 export * from './ccd-tab-fields.pipe';
-export * from './cdd-page-fields.pipe';
+export * from './ccd-page-fields.pipe';
 export * from './fields-filter.pipe';
